requirements.txt
flask==1.*
bson==0.*
pymongo==3.*
mysql-connector-python==8.*
pandas==0.*
numpy==1.*
PyJWT==1.*
pytest==5.*
coverage==5.*